requirements.txt
sentence_transformers
pdfminer.six
nltk
numpy